+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/DoubleStorage.java
@@ -8,6 +8,7 @@ import org.enso.table.data.column.operation.map.UnaryMapOperation;
 import org.enso.table.data.column.operation.map.numeric.DoubleBooleanOp;
 import org.enso.table.data.column.operation.map.numeric.DoubleIsInOp;
 import org.enso.table.data.column.operation.map.numeric.DoubleNumericOp;
+import org.enso.table.data.column.operation.map.UnaryDoubleToLongOp;
 import org.enso.table.data.column.storage.BoolStorage;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.type.FloatType;
@@ -242,6 +243,27 @@ public final class DoubleStorage extends NumericStorage<Double> {
               protected double doDouble(
                   double a, double b, int ix, MapOperationProblemBuilder problemBuilder) {
                 return Math.pow(a, b);
+              }
+            })
+        .add(
+            new UnaryDoubleToLongOp(Maps.TRUNCATE) {
+              @Override
+              protected long doOperation(double a) {
+                return (long) a;
+              }
+            })
+        .add(
+            new UnaryDoubleToLongOp(Maps.CEIL) {
+              @Override
+              protected long doOperation(double a) {
+                return (long) Math.ceil(a);
+              }
+            })
+        .add(
+            new UnaryDoubleToLongOp(Maps.FLOOR) {
+              @Override
+              protected long doOperation(double a) {
+                return (long) Math.floor(a);
               }
             })
         .add(