+++ CHANGELOG.md
@@ -479,6 +479,7 @@
 - [Handling edge cases in rounding.][6922]
 - [Split `Table.create_database_table` into `Connection.create_table` and
   `Table.select_into_database_table`.][6925]
+- [Speed improvements to `Column` `.truncate`, `.ceil`, and `.floor`.][6941]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -696,6 +697,7 @@
 [6890]: https://github.com/enso-org/enso/pull/6890
 [6922]: https://github.com/enso-org/enso/pull/6922
 [6925]: https://github.com/enso-org/enso/pull/6925
+[6941]: https://github.com/enso-org/enso/pull/6941
 
 #### Enso Compiler